updated to latest sources

+++ app/src/main/java/com/eveningoutpost/dexdrip/cgm/carelinkfollow/message/ActiveInsulin.java
@@ -2,14 +2,11 @@ package com.eveningoutpost.dexdrip.cgm.carelinkfollow.message;
 
 import java.util.Date;
 
-/**
- * CareLink ActiveInsulin data
- */
 public class ActiveInsulin {
 
     public Integer code;
     public Date datetime;
-    public long version;
+    public int version;
     public Double amount;
     public String precision;
     public String kind;
